Update the GitHub link in case #234
The link to the immortal hack in Apollo 11's source is currently broken. This may be a bit of a fool's errand, seeing as people are still contributing erratas to the transcription, and the link may change more in the coming month, but no harm in trying to keep it as up-to-date as we can.